requirements.txt
weaviate-client
spotipy==2.23.0
streamlit==1.25.0
librosa==0.10.0.post2
numpy==1.24.4
scikit-learn==1.2.2
transformers==4.35.0
